drf_dvd/requirement.txt
django>=4.0.0,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers
psycopg2==2.9.9